Fix wonkiness in text editor commands
TextEditorCommands are not allowed to be async - the editBuilder they are provided is applied immediately without waiting. This is why some of this code was creating its own new EditBuilders (without it, their edits were lost). To avoid confusion, these async operations are now all standard commands and read the active text editor themselves.

Fixes #734.