Track pending resumes so we don't send too many
This feels a bit dirty, but the current code allowed sending multiple resumes, eg. if we got an IsolateRunnable event at startup right after we sent (but had not had a response to) a Resume. This code currently just silently drops resumes if one is in progress (or the thread is not paused) which feels bad, but better than sending them and potentially unpausing later (when unexpected).

Probably there's a better way to do it, and we should consider a bigger refactor of some of this once the test coverage is better.

Fixes #902. Probably. Maybe. Hopefully.