Unskip tests and handle rename ranges better
This is a hack - VS Code is (pointlessly) validating that the rename range we provide spans the position. This is not the case when we want to rename from the "class" keyword, for example (or imports).

Since it doesn't actually validate the provided placeHolder is inside that range, we can just give it back the position it gave us, even though the placeholder isn't there (it does mean the input box doesn't get moved, but that's fairly minor).

Fixes #910.