Handle keeping results across multiple files
flutter run doesn't support multiple suites with --start-paused so we staash results by the suite path so you can run multiple. This code somewhat allows for multiple-suites within a "suite" just in case we somehow end up with many suites back from a single run (but if this becomes normal, we may wish to adjust how it displays).