Split watch/evaluate tests
The first one was expecting the *same* results but was doing child-variables, which means we got mismatches when the child had a truncated value (which would be a full value when evaluated).

Now, the original test just checks top-level variables, and the new tests checks evaluates, but takes into account truncation.