Fix flaky test caused by "top 500" results being truncated differently based on order
"My" matches a lot of things (since we do fuxxy matches like "*m*y*"), so the flakes appear to be when the test items were not within the top 500 (which could be based on timing in the analysis server - they're not sorted). Now everything is prefixed with "MyTest" to make the search more specific.