Revert sorting of document changes
Reverts 725583385fc6afb9525b82a77d8921c0c77a82fd.

The LSP docs state (and it's assumed the VS Code API is intended to be the same) that edits should always be applied in the order they are sent. While this works with VS Code today, it's not guaranteed, as it would be valid (according to the spec) to send edits that do not work in offset-descending order because the offsets already take into account the other edits.