Don't require a VM Service connection to enable hot reload/restart for Flutter
In the new SDK DAPs we don't get a VM Service connection when running without debugging. Hot Reload and Hot Restart do not require calling the services directly, so do not need to be checked in this way.

Additionally, `flutterHotRestartOnSave` was for legacy web support and no longer used, since web advertises Hot Reload (even though it just Hot Restarts) so was removed (it's no longer possible to perform this check without a VM Service connection).