Don't require a first path segment to be able to forceFlutterDebug
The change in 74fd1ff0b6968cd59d3a03d2472897c25de482de tried to avoid skipping the Flutter debugger for tests inside the "tool" folder but accidentally skipped using the Flutter debugger for anything where we couldn't compute the first path segment.

This change means if we can't compute a first path segment (which we can't for an internal `//` target) we can still check for a Flutter project or forceFlutterDebug.