Show daemons logs in an Output panel and auto-show on certain types of errors
We had some handling of "daemon.log" here but it seems unused and the errors we care about here are from "daemon.logMessage", so now we handle both.

We had previously (e504fb4b2f436b905b4466166764d652310ca159) stopped showing this automatically because it was noisy, so now we do it only when the output contains some specific strings that we know are errors the user may be interested in.

Fixes https://github.com/Dart-Code/Dart-Code/issues/4839