Ensure stdout buffer is always flushed
A failing test is caused by this never being flushed because the output does not have a newline. The missing newline may be a bug (see https://github.com/dart-lang/sdk/issues/54582) but we should never lose the last line of stdout even without a newline.

This issue only affects the legacy DAPs, as the new DAPs don't buffer like this.