Limit shell execute only to .bat on Windows
Seems like we only need shell execute and special escaping for paths with particular characters in. For most things, just spawn() directly with unescaped args works fine (and it always does for non-Windows).

By eliminating shell execute, we reduce the chance of leaking processes where we send SIGTERM to the shell and it doesn't get passed on to the spawned process.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5018