Force terminate the analyzer process when disposing
This may fix https://github.com/Dart-Code/Dart-Code/issues/5084 which I suspect is because the server is busy and doesn't handle the graceful shutdown (and we don't appear to have anything to try to terminate the process directly).

I'm unable to reproduce on Mac or Windows, so I don't yet know if this will help.