Add the ability to show notifications when the current SDK is unsupported (or will be soon)
This adds two new fields to `DartCapabilities` for `isUnsupportedNow` and `isUnsupportedSoon` to determine whether to warn the user about unsupported SDKs.

We always do the SDK check against the Dart SDK version, but if the current Dart SDK comes from Flutter, we will reference "Flutter" and its SDK version in the user-facing messages.

"Unsupported now" messages will always be shown as the intention is that the user will either upgrade their SDK or switch to an older extension.. this may seem annoying, but the extension will be broken in many ways when using an SDK that triggers this, so ignoring it is not an option - you must upgrade SDK or switch to an older extension.

"Unsupported soon" messages will be shown only until dismissed the first time (clicking "More Info" counts as dismissing), but this is tracked by major+minor versions, so dismissing for Dart 2.0 would not prevent seeing the message again in future for 3.0.