Refactor packages view so the tree data provide is no longer coupled directly to the "pub deps --json" format
The way we're using this format doesn't work well with Workspaces, so this refactors things so the `PubDeps` class can convert all of the JSON data into a complete tree structure that can be used by the packages view data provider more cleanly.

This does not add support for Pub Workspaces, but it should make doing so much simpler.