Force allThreadsContinued to false in DAP middleware
This is a quick-fix for https://github.com/Dart-Code/Dart-Code/issues/5252 because there will be a delay in shipping a fix in the debug adapter.

The default for `allThreadsContinued` is not documented well... The docs make it seem like `false` is the default, but VS Code treats it like `true`. This makes it unambiguous (we never use it for all-threads, but this code only sets it if it is `undefined` to still allow for the possibility of an explicit `true`).

Fixes https://github.com/Dart-Code/Dart-Code/issues/5252