Merge all test output for failed() summary
For Flutter unit tests the output was all in a single event so we just showed the last output event. However, for integration tests the lines are split up.

I can't find any cases where we wouldn't want to show the whole output, so this joins all the events together and uses them.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5353