Use workspace file folder for relative paths when in a workspace
Previously, when resolving a relative path in something like an SDK or log file path, we would use the first open folder as the root path. This was very arbitrary and might not be the obvious thing if you're set the setting at the workspace level.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5401.