Sort test results by label for tests that modify offsets
This test re-runs tests after modifying the document which results in inconsistent ordering (after the first test was re-run after the insertion, but before the second was re-run).

Also only check results at the end, otherwise the duplicated tests can have different results (because the first instance of each passes, and the second fails).