Ensure we clear out the TestRun for a debug session when it ends
Because restarts can reuse the same session ID (and we can't change this), we need to be sure we don't try to run an ended test run to report results after the restart. This removes the TestRun from the mapping when a session ends, so if another session starts using the same ID, it will produce a new TestRun (and therefore be recorded separately in the history).

Fixes https://github.com/Dart-Code/Dart-Code/issues/5481