Allow launch config to force a specific debugger type
Generally users shouldn't use this themselves in `launch.json` (because we should correctly infer it in most cases - and if we don't, we should improve the heuristics). However in the case of attaching to an arbitrary VM Service regardless of the open project (for example from another extension or an autolaunch file) I think it's fair to want to force this.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5535