Keep the "Starting debug session…" notification visible longer
We used to hide this on any message, but for integration test runs we get some basic messages early on, but then nothing for a long time while the app builds.

With this change, we'll only hide this notification when there is an output event, or a progress event from the debug adapter.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5654