Ping over the DTD WebSocket every 15s to avoid some proxies (like Norton) dropping the connection
Users with Norton 360 installed found that if there is no traffic over the WebSocket for 60s, it gets dropped. This appears to be that the connections are proxied (even if all scanning is disabled) and "idle" connections are dropped.

This forces traffic ever 15s which will prevent them from dropping the connection.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5794