Fix test tracking when ranges are deleted
When processing changes, we were firing the callbacks in-line while enumerating the trackers. A deleted position could result in a tracked range from disposing itself (the start position is deleted first, and it then disposes and untracks the end position) which modified the list of trackers while we were enumerating.

To fix, we store up the callbacks so nothing is modified while we synchronously process the changes.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5831