Revert to stable vscode-languageclient v9
The feature we switched to this for (`delayOpenNotifications`) is broken so was previously disabled. As it's unclear if this feature will be fixed (and not just removed), switch back to a stable version to reduce risk.

(the vNext was only ever shipped in a pre-release version of the extension, so this avoids it going to stable)

See https://github.com/microsoft/vscode-languageserver-node/issues/1695
See https://github.com/dart-lang/sdk/issues/62209